Extract workspace composer state hook
